packages/api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
qdrant-client>=1.12.0
openai>=1.50.0
numpy>=1.26.0
PyJWT>=2.8.0
